test: close cross-language gaps; fix: explicit fractional-sqm rounding rule (spec 1.1)
- Spec Sec 1.1: base = round_half_up(construction_cost_per_sqm_pence x
  total_construction_sqm), rounded once at source before contingency. Applied at
  the single multiply site in both engines (calculateTotalConstructionCost /
  calculate_total_construction_cost). TDD RED-first regressions in both languages
  (round_half_up(50000 x 500.5) = 25,025,000; the odd-half round_half_up(333 x
  100.5) = 33,467). Integer-sqm fixtures are unaffected (rounding an integer
  product is identity).
- Ported the TS 4-way derived-variant invariant matrix (base/retain_all/serviced/
  term=1) into Python: TestInvariantMatrix in test_financial_model_fixtures.py,
  one test method per TS it() (7), parametrised over the same 3-fixture x
  4-variant = 12-way matrix (84 checks total, matching TS's assertion-group
  count). The original test_invariants is kept alongside, not superseded -- it's
  a stronger unconditional check over the base fixtures.
- Ported the 4 hand-derived v1->v2 migration cases from migrate.test.ts into
  tests/test_migrate_v2.py, sibling to test_migrate_v3.py's existing v2->v3 port.
- Minor fixes folded in: AppraisalResultV2.lender_gdv_pence/lender_gdv_variance_pence
  now float | None (Task 3 made non-integer values possible for fixed_amount/
  per_unit lender-valuation bases); corrected the breakeven.test.ts /
  test_financial_model_breakeven.py 10**80 iteration-cap comments, which
  described Python's arbitrary-precision mechanism for both languages -- JS
  numbers are IEEE-754 doubles and hit the cap via float-stall (~52 real
  halvings then adjacent-double stalling), not ~266 genuine bisection steps.
- docs/financial-model/test-cases.md Sec 4/Sec 7 updated: both recorded
  cross-language gaps are closed; the Sec 1.1 rounding regressions are
  registered in the rounding-parity entry.

+++ app/financial_model/metrics.py
@@ -97,8 +97,11 @@ def solve_irr(cashflows: list[float]) -> float | None:
 class AppraisalResultV2:
     calc_version: str
     gdv_pence: int
-    lender_gdv_pence: int | None
-    lender_gdv_variance_pence: int | None
+    # float, not int (Task 3's fix made non-integer lender_gdv_pence possible for
+    # fixed_amount/per_unit bases -- see LenderGdvResult.lender_gdv_pence in
+    # lender_valuation.py; variance inherits the same fractional possibility).
+    lender_gdv_pence: float | None
+    lender_gdv_variance_pence: float | None
     lender_gdv_variance_pct: float | None
     acquisition_cost_pence: int
     sdlt_pence: int
